refactor(parser): rename classes to better show their purpose
