+++ package-parser/package_parser/cli/_run_api.py
@@ -5,7 +5,7 @@ from typing import Optional
 from package_parser.cli._json_encoder import CustomEncoder
 from package_parser.cli._shared_constants import _API_KEY
 from package_parser.processing.api import get_api
-from package_parser.processing.api.documentation_parsing import DocstringStyle
+from package_parser.processing.api.docstring_parsing import DocstringStyle
 from package_parser.processing.api.model import API
 from package_parser.processing.dependencies import get_dependencies
 from package_parser.utils import ensure_file_exists
