+++ package-parser/package_parser/processing/api/_ast_visitor.py
@@ -17,14 +17,14 @@ from package_parser.utils import parent_qualified_name
 
 from ._file_filters import _is_init_file
 from ._get_parameter_list import get_parameter_list
-from .documentation_parsing import AbstractDocumentationParser
+from .docstring_parsing import AbstractDocstringParser
 
 
 class _AstVisitor:
     def __init__(
-        self, documentation_parser: AbstractDocumentationParser, api: API
+        self, docstring_parser: AbstractDocstringParser, api: API
     ) -> None:
-        self.documentation_parser: AbstractDocumentationParser = documentation_parser
+        self.docstring_parser: AbstractDocstringParser = docstring_parser
         self.reexported: dict[str, list[str]] = {}
         self.api: API = api
         self.__declaration_stack: list[Union[Module, Class, Function]] = []
@@ -148,7 +148,7 @@ class _AstVisitor:
             superclasses=class_node.basenames,
             is_public=self.is_public(class_node.name, qname),
             reexported_by=self.reexported.get(qname, []),
-            documentation=self.documentation_parser.get_class_documentation(class_node),
+            documentation=self.docstring_parser.get_class_documentation(class_node),
         )
         self.__declaration_stack.append(class_)
 
@@ -181,7 +181,7 @@ class _AstVisitor:
             qname=qname,
             decorators=decorator_names,
             parameters=get_parameter_list(
-                self.documentation_parser,
+                self.docstring_parser,
                 function_node,
                 self.__get_id(function_node.name),
                 qname,
@@ -190,7 +190,7 @@ class _AstVisitor:
             results=[],  # TODO: results
             is_public=is_public,
             reexported_by=self.reexported.get(qname, []),
-            documentation=self.documentation_parser.get_function_documentation(
+            documentation=self.docstring_parser.get_function_documentation(
                 function_node
             ),
         )
