+++ package-parser/package_parser/processing/api/docstring_parsing/_DocstringStyle.py
